# docs/swagger.yaml
basePath: /api/v1
definitions:
  POS-kasir_internal_common.ErrorResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
  POS-kasir_internal_common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  POS-kasir_internal_dto.CancellationReportResponse:
    properties:
      cancelled_orders:
        type: integer
      reason:
        type: string
      reason_id:
        type: integer
    type: object
...
  POS-kasir_internal_dto.CashierPerformanceResponse:
    properties:
      order_count:
        type: integer
      total_sales:
        type: number
      user_id:
        type: string
      username:
        type: string
    type: object
...
  POS-kasir_internal_dto.DashboardSummaryResponse:
    properties:
      total_orders:
        type: integer
      total_products:
        type: integer
      total_sales:
        type: number
      unique_cashier:
        type: integer
    type: object
...
  POS-kasir_internal_dto.PaymentMethodPerformanceResponse:
    properties:
      order_count:
        type: integer
      payment_method_id:
        type: integer
      payment_method_name:
        type: string
      total_sales:
        type: number
    type: object
...
  POS-kasir_internal_dto.ProductPerformanceResponse:
    properties:
      product_id:
        type: string
      product_name:
        type: string
      total_quantity:
        type: integer
      total_revenue:
        type: number
    type: object
  POS-kasir_internal_dto.ProductProfitResponse:
    properties:
      gross_profit:
        type: number
      product_id:
        type: string
      product_name:
        type: string
      total_cogs:
        type: number
      total_revenue:
        type: number
      total_sold:
        type: integer
    type: object
...
  POS-kasir_internal_dto.ProfitSummaryResponse:
    properties:
      date:
        type: string
      gross_profit:
        type: number
      total_cogs:
        type: number
      total_revenue:
        type: number
    type: object
...
  POS-kasir_internal_dto.SalesReport:
    properties:
      date:
        type: string
      order_count:
        type: integer
      total_sales:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: POS Kasir API
  title: POS Kasir API
  version: "1.0"
paths:
...
  /reports/cancellations:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/POS-kasir_internal_dto.CancellationReportResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get cancellation reports
      tags:
      - Reports
  /reports/cashier-performance:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/POS-kasir_internal_dto.CashierPerformanceResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get cashier performance
      tags:
      - Reports
  /reports/dashboard-summary:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.DashboardSummaryResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get dashboard summary
      tags:
      - Reports
  /reports/payment-methods:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/POS-kasir_internal_dto.PaymentMethodPerformanceResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get payment method performance
      tags:
      - Reports
  /reports/products:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/POS-kasir_internal_dto.ProductPerformanceResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get product performance
      tags:
      - Reports
  /reports/profit-products:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/POS-kasir_internal_dto.ProductProfitResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get product profit reports
      tags:
      - Reports
  /reports/profit-summary:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/POS-kasir_internal_dto.ProfitSummaryResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get profit summary
      tags:
      - Reports
  /reports/sales:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/POS-kasir_internal_dto.SalesReport'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get sales reports
      tags:
      - Reports
...
swagger: "2.0"